Move RealizationStorageState into storage module
