+++ src/ert/libres_facade.py
@@ -23,9 +23,9 @@ from ert.config import (
 )
 from ert.data import MeasuredData
 from ert.data._measured_data import ObservationError, ResponseError
-from ert.realization_storage_state import RealizationStorageState
 from ert.shared.version import __version__
 from ert.storage import EnsembleReader
+from ert.storage.realization_storage_state import RealizationStorageState
 
 from .analysis._es_update import UpdateSettings
 from .enkf_main import EnKFMain, ensemble_context
