+++ src/ert/run_models/ensemble_experiment.py
@@ -8,9 +8,9 @@ import numpy as np
 
 from ert.enkf_main import sample_prior
 from ert.ensemble_evaluator import EvaluatorServerConfig
-from ert.realization_storage_state import RealizationStorageState
 from ert.run_context import RunContext
 from ert.storage import EnsembleAccessor, StorageAccessor
+from ert.storage.realization_storage_state import RealizationStorageState
 
 from .base_run_model import BaseRunModel
