+++ src/ert/shared/hook_implementations/workflows/export_misfit_data.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from ert import ErtScript
 from ert.exceptions import StorageError
-from ert.realization_storage_state import RealizationStorageState
+from ert.storage.realization_storage_state import RealizationStorageState
 
 
 class ExportMisfitDataJob(ErtScript):
