+++ src/ert/storage/local_storage.py
@@ -28,10 +28,10 @@ from filelock import FileLock, Timeout
 from pydantic import BaseModel, Field
 
 from ert.config import ErtConfig
-from ert.realization_storage_state import RealizationStorageState
 from ert.shared import __version__
 from ert.storage.local_ensemble import LocalEnsembleAccessor, LocalEnsembleReader
 from ert.storage.local_experiment import LocalExperimentAccessor, LocalExperimentReader
+from ert.storage.realization_storage_state import RealizationStorageState
 
 if sys.version_info < (3, 11):
     from typing_extensions import Self
