+++ src/ert/storage/realization_storage_state.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class RealizationStorageState(Enum):
+    UNDEFINED = 1
+    INITIALIZED = 2
+    HAS_DATA = 4
+    LOAD_FAILURE = 8
+    PARENT_FAILURE = 16
