+++ tests/performance_tests/test_memory_usage.py
@@ -12,8 +12,8 @@ from flaky import flaky
 from ert.analysis import UpdateConfiguration, smoother_update
 from ert.config import ErtConfig, SummaryConfig
 from ert.enkf_main import sample_prior
-from ert.realization_storage_state import RealizationStorageState
 from ert.storage import open_storage
+from ert.storage.realization_storage_state import RealizationStorageState
 from tests.performance_tests.performance_utils import make_poly_example
