+++ tests/unit_tests/analysis/test_es_update.py
@@ -26,8 +26,8 @@ from ert.cli import ENSEMBLE_SMOOTHER_MODE
 from ert.cli.main import run_cli
 from ert.config import AnalysisConfig, ErtConfig, GenDataConfig, GenKwConfig
 from ert.config.analysis_module import ESSettings, IESSettings
-from ert.realization_storage_state import RealizationStorageState
 from ert.storage import open_storage
+from ert.storage.realization_storage_state import RealizationStorageState
 
 
 @pytest.fixture
