+++ tests/unit_tests/test_load_forward_model.py
@@ -12,8 +12,8 @@ from resdata.summary import Summary
 from ert.config import ErtConfig
 from ert.enkf_main import create_run_path, ensemble_context
 from ert.libres_facade import LibresFacade
-from ert.realization_storage_state import RealizationStorageState
 from ert.storage import open_storage
+from ert.storage.realization_storage_state import RealizationStorageState
 
 
 @pytest.fixture()
